+++ src/elements/images.rs
@@ -5,9 +5,9 @@ use crate::PelicanUI;
 pub struct Icon;
 
 impl Icon {
-    pub fn new(ctx: &mut Context, name: &'static str, color: Color, size: u32) -> Image {
+    pub fn new(ctx: &mut Context, name: &'static str, color: Color, size: f32) -> Image {
         let icon = ctx.get::<PelicanUI>().theme.icons.get(name);
-        Image{shape: ShapeType::Rectangle(0, (size, size)), image: icon, color: Some(color)}
+        Image{shape: ShapeType::Rectangle(0.0, (size, size)), image: icon, color: Some(color)}
     }
 }
 impl Events for Icon {}
@@ -16,8 +16,8 @@ impl Events for Icon {}
 pub struct Brand;
 
 impl Brand {
-    pub fn new(image: resources::Image, size: (u32, u32)) -> Image {
-        Image{shape: ShapeType::Rectangle(0, (size.0, size.1)), image, color: None}
+    pub fn new(image: resources::Image, size: (f32, f32)) -> Image {
+        Image{shape: ShapeType::Rectangle(0.0, (size.0, size.1)), image, color: None}
     }
 }
 impl Events for Brand {}
